requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests==2.26.0
python-dotenv==1.0.0
orator==0.9.9
psycopg2-binary==2.9.6
pylint==2.17.4
autopep8==2.0.2
pytest==7.3.2
pytest-cov==4.1.0
pytest-mock==3.11.1
factory_boy==3.2.1
python-jose==3.3.0
bcrypt==4.0.1